+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Although Moulti's Python packages, modules and functions are obviously available
 - `Ctrl+t` / `Ctrl+y` jumps to the previous/next unanswered question
 - `O` / `X` collapses/expands new steps
 - the environment variable `MOULTI_ENFORCE_COLLAPSIBLE` does the same as soon as Moulti starts
+- `Ctrl+o` / `Ctrl+x` collapses/expands both existing and new steps
 
 ## [1.14.0] - 2024-08-04